Relocates RequestProcessor and ResponseHandler into the Java library.
