+++ bluebotruntimelib/src/test/java/com/bluebot/ui/RequestProcessorSpec.java
@@ -1,0 +1,16 @@
+package com.bluebot.ui;
+
+import org.junit.Test;
+
+/**
+ * Created by Clifton Craig on 4/11/17.
+ * Copyright GE 4/11/17
+ */
+
+public class RequestProcessorSpec {
+    @Test
+    public void shouldPassRequestsToRelevantCode() throws Exception {
+
+
+    }
+}
